backend/requirements.txt
django
djangorestframework
django-simple-history
django-extensions
djangorestframework-simplejwt
django-debug-toolbar
django-storages
boto3
pillow
django-widget-tweaks
django-cors-headers
django-filter
requests
django-celery-beat
django-celery-results
celery
redis
django-redis
